+++ base/src/main/java/proguard/analysis/cpa/jvm/domain/value/JvmValueBamCpaRun.java
@@ -151,7 +151,7 @@ public class JvmValueBamCpaRun
                 cfa,
                 mainSignature,
                 valueFactory,
-                new ExecutingInvocationUnit(valueFactory),
+                new ExecutingInvocationUnit.Builder().setEnableSameInstanceIdApproximation(true).build(valueFactory),
                 maxCallStackDepth,
                 heapModel,
                 staticFields,